Refactor VI: Define packages' dependencies
