+++ Code/Language/m/src/m/validation/global/GlobalData.java
@@ -5,16 +5,15 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-import org.eclipse.lsp4j.Diagnostic;
-
 import m.library.types.Type;
 import m.model.Function;
+import m.validation.problems.Problem;
 
 public class GlobalData
 {
 	public Map<String,Type> components;
 	public List<Function> functions;
-	public Map<String,List<Diagnostic>> diagnostics;
+	public Map<String,List<Problem>> diagnostics;
 
 	public GlobalData()
 	{
