+++ Code/Language/m/src/m/validation/problems/errors/UndefinedSymbol.java
@@ -22,7 +22,6 @@ public class UndefinedSymbol extends Problem
 		this.feature = feature;
 	}
 
-	@Override
 	public List<Diagnostic> diagnostics(Library library, String text)
 	{
 		var list = new ArrayList<Diagnostic>();
